% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expect_called.R
\name{expect_called}
\alias{expect_called}
\title{Expect a given function to be called during an expression.}
\usage{
expect_called(function_name, expr)
}
\arguments{
\item{function_name}{name. Either the name of the function, or
a package function given by the \code{::} or \code{:::} operators.}

\item{expr}{expression. An R expression that should somewhere during
it execution call \code{function_name}.}
}
\value{
\code{NULL}, invisibly, but an error if the expectation
   is not satisfied (i.e., \code{function_name} is not called).
}
\description{
Expect a given function to be called during an expression.
}
\examples{
expect_called(force, force(1))
}

