% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/let.R
\name{let}
\alias{let}
\title{A mocha-style let helper for easy object mocking.}
\usage{
let(name, expr, envir = parent.frame())
}
\arguments{
\item{name}{character. The name of the mocked object.}

\item{expr}{expression. The expression to re-evaluate each time the
mock is instantiated.}

\item{envir}{environment. The environment to create an active binding
in which will create mocked locals (by default the calling environment).}
}
\description{
A mocha-style let helper for easy object mocking.
}

