% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make an Analysis Project}
\usage{
make_project(path)
}
\arguments{
\item{path}{Path automatically sent by research_project.dcf}
}
\value{
a project
}
\description{
This function makes a R project that includes an analysis.Rmd file with 
conflicted and tidyverse.  This is used by the research_project.dcf file.
}
