% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rUM-package.R
\docType{package}
\name{rUM-package}
\alias{rUM}
\alias{rUM-package}
\title{rUM: R Templates from the University of Miami}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This holds r markdown and quarto templates for academic papers and slide decks. It also has templates to create research projects which contain academic papers as vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://raymondbalise.github.io/rUM/}
  \item \url{https://github.com/RaymondBalise/rUM}
  \item Report bugs at \url{https://github.com/RaymondBalise/rUM/issues}
}

}
\author{
\strong{Maintainer}: Raymond Balise \email{balise@miami.edu} (\href{https://orcid.org/0000-0002-9856-5901}{ORCID})

Authors:
\itemize{
  \item Gabriel Odom \email{gabriel.odom@fiu.edu} (\href{https://orcid.org/0000-0003-1341-4555}{ORCID})
  \item Kyle Grealis \email{kxg679@miami.edu} (\href{https://orcid.org/0000-0002-9223-8854}{ORCID})
  \item Francisco Cardozo \email{foc9@miami.edu} (\href{https://orcid.org/0000-0002-1925-4954}{ORCID})
}

Other contributors:
\itemize{
  \item Frank Gutierrez \email{frank14898@outlook.com} (\href{https://orcid.org/0009-0005-9742-6992}{ORCID}) [contributor]
}

}
\keyword{internal}
