% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rum-package.R
\docType{package}
\name{rum}
\alias{rum}
\alias{rum-package}
\title{Stubbing and mocking for R.}
\description{
The rum packages aims to provide extra testing capabilities.
Specifically, functions like \code{expect_receives} and
\code{expect_called} are introduced that ensure a given
function is called, possibly with specific parameters.
}
\details{
This is very useful if you wish to, for example, ensure that
a certain function that makes a network request was called,
but you do not want to perform messy stubbing that would
achieve the same effect.
}
\seealso{
\url{https://github.com/freerange/mocha}
}

