% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scss.R
\name{write_scss}
\alias{write_scss}
\title{Create a Quarto SCSS file}
\usage{
write_scss(name = "custom", path = getwd())
}
\arguments{
\item{name}{The name of the scss file without extension. Default \code{name} is 
"custom".}

\item{path}{The path to the main project level. Defaults to the current
working directory}
}
\value{
A \code{.scss} file to customize Quarto styling.
}
\description{
This function creates the \code{.scss} file so that any Quarto project can be easily 
customized with SCSS styling variables, mixins, and rules.
}
\details{
For more information on customizing Quarto documents with SCSS, please refer to
\url{https://quarto.org/docs/output-formats/html-themes.html#customizing-themes},
\url{https://quarto.org/docs/output-formats/html-themes-more.html}, and 
\url{https://github.com/twbs/bootstrap/blob/main/scss/_variables.scss} will provide
you with over 1500 lines of SCSS variables.
}
\examples{
\dontrun{
write_scss(name = "custom", path = "path/to/project")
}
}
