% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expect_receives.R
\name{expect_receives}
\alias{expect_receives}
\title{Expect a given function to be receive specified arguments during execution.}
\usage{
expect_receives(function_name, arguments, expr, times = 1)
}
\arguments{
\item{function_name}{}

\item{expr}{expression. An R expression that should somewhere during
it execution call \code{function_name}.}
}
\description{
This assertion checks that a given function is called with specified
arguments during the execution of another expression.
}

