% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_slides.R
\name{find_slides}
\alias{find_slides}
\title{List all slide decks in a package}
\usage{
find_slides(package = NULL)
}
\arguments{
\item{package}{Character. Provide the package containing one or more slide decks.}
}
\value{
A list of class "slide_finder" containing the name of the package and the
name of the slides.
}
\description{
Returns the names of all Quarto slide decks in a package. This is designed to work
with \code{rUM::show_slides()} to preview the slide deck. For more
information look in the
\href{../doc/ah_write_slides.html}{Creating Slides with write_slides()} vignette.
}
\examples{
if (interactive()) {
  find_slides("rUM")
}

}
