% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make an Analysis Project}
\usage{
make_project(path)
}
\arguments{
\item{path}{Path automatically sent by research_project.dcf}
}
\value{
Creates a project directory with stuff in it.
}
\description{
This function makes an R project that includes an analysis.Rmd 
file with conflicted and tidyverse and an aggressive .gitignore.  The 
.gitignore is designed to help protect against leaking data (with protected 
health information). This function is used by the research_project.dcf file 
to make the files.
}
