% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_man.R
\name{write_man}
\alias{write_man}
\title{Write a manual page for package dataset documentation}
\usage{
write_man(the_dataset)
}
\arguments{
\item{the_dataset}{Dataset object}
}
\value{
A \code{.Rd} file in the \code{man} package directory corresponding to the 
supplied dataset.
}
\description{
This function will parse through a dataset and create a manual page. The dataset
documentation will output a Markdown-style table with the type, factor levels (if 
appropriate), the range of numeric values, and a generic description section. 
TODO: ADD MORE CONTENT!
}
\examples{
if (interactive()) write_man(mtcars)
}
