% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_readme.R
\name{write_readme}
\alias{write_readme}
\title{Create a project README file}
\usage{
write_readme(path = getwd())
}
\arguments{
\item{path}{The path to the main project level. Defaults to the current
working directory.}
}
\value{
A README.md template. Contains sections for project description (study
name, principle investigator, & author), project setup steps for ease of portability,
project file descriptions, project directory descriptions, and miscellaneous.
}
\description{
This function is designed to first check for the existence of a README file. If none
is detected or the user chooses to overwrite the current README, a README template
is loaded from \code{rUM/inst/gists/README.md} and written to the project level.
}
\details{
This function handles the creation and/or overwriting of both README.md
and dated_progress_notes.md files. For each file, if it already exists, the user
will be prompted whether to overwrite it. The dated_progress_notes.md file will
be initialized with the current date and a "project started" message.

NOTE: Some documentation remains to provide the user with example descriptions for 
files & directories. It is highly recommended to keep these sections. This is a 
modifiable template and should be tailor-fit for your exact purpose.
}
\examples{
\dontrun{
write_readme(path = "path/to/project")
}
}
